Adding quaternion.h

+++ include/Matrix3x3.h
@@ -6,11 +6,11 @@
 class Quaternion;
 class Vector3D;
 
-class Matrix3x3 : public Matrix<double,3,3>
+class Matrix3x3 : public Matrix<float,3,3>
 {
 public:
     Matrix3x3();
-    Matrix3x3(std::initializer_list<std::initializer_list<double>>);
+    Matrix3x3(std::initializer_list<std::initializer_list<float>>);
 
     Matrix3x3 operator*(const Matrix3x3& other) const;
     Matrix3x3 operator*(const Vector3D& other) const;
