+++ src/Matrix3x3.cpp
@@ -7,7 +7,7 @@ Matrix3x3::Matrix3x3()
     : Matrix<double,3,3>()
 { }
 
-Matrix3x3::Matrix3x3(std::initializer_list<std::initializer_list<double>> row_list)
+Matrix3x3::Matrix3x3(std::initializer_list<std::initializer_list<float>> row_list)
     : Matrix<double,3,3>(row_list)
 { }
